Read big picture window title from localization files
Rather than hardcoding the steam big picture mode window title for each language
into the source code, it's better to read it from steam's localization files
in the steam install directory. This ensures that we use the correct window
title regardless of the steam version installed. It also means we do not have
to update the window titles in this repository if steam adds support for more
languages.