Adjust teardown script to ensure steam window is properly closed
It seems steam has changed some of the behavior with how the steam window is opened
upon closing big picture mode. To ensure the window is closed properly, this commit
requires more consecutive instances of the window reporting as closed. This means
there should be less instances of the window being left open after the stream ends.